Fixed issue with the REST API because the ID fields weren't the correct type. They need to by Long not Integer!

+++ src/main/java/uk/org/scbc/entities/Building.java
@@ -8,9 +8,10 @@ import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 
 /**
- * Created by marc.thomas on 20/04/2016.
+ * Buidling Entity
+ *
+ * @author marc.thomas
  */
-
 @Entity
 @Data
 @NoArgsConstructor
@@ -19,7 +20,7 @@ import javax.validation.constraints.NotNull;
 public class Building {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     @NotNull
     private String name;
